move user remote functionality to ee
